Various small fixes (again).
	* mln/core/image/complex_image.hh: update comments.

	* mln/core/image/complex_neighborhood_piter.hh: avoid a warning.

	* mln/core/internal/site_relative_iterator_base.hh: call center_at_()
	earlier.

	* mln/core/site_set/p_vertices.hh: call the proper graph() method.

	* mln/fun/v2v/ch_function_value.hh,
	* mln/trait/ch_function_value.hh: fix a recursive inclusion.

	* tests/core/image/graph_image.cc: fix layout.

	* tests/core/image/line_graph_image.cc: cleanup includes.

	* tests/level/paste_full.cc: move to...
	* tests/data/paste_full.cc: ... this file.

	* tests/level/Makefile.am: remove obsolete entries.

	* tests/histo/to_image1d.cc: Fix compilation issue. An incorrect image
	type was expected.

	* tools/area_flooding.cc: fix a warning.

	* tests/core/site_set/p_vertices.cc: The function mapping vertices to
	sites was not large enough to hold all the values.

	* mln/fun/x2v/bilinear.hh: avoid warnings using static_cast.

git-svn-id: https://svn.lrde.epita.fr/svn/oln/trunk@3122 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ milena/mln/fun/x2v/bilinear.hh
@@ -98,10 +98,10 @@ namespace mln
         double y2 = std::floor(v[1]) + 1;
 
         //if (not ima.has(point2d(x1, y1))) return mln_value(I)();
-        vsum q11 = ima(point2d(x1, y1));
-        vsum q12 = ima(point2d(x1, y2));
-        vsum q21 = ima(point2d(x2, y1));
-        vsum q22 = ima(point2d(x2, y2));
+        vsum q11 = ima(point2d(static_cast<unsigned>(x1), static_cast<unsigned>(y1)));
+        vsum q12 = ima(point2d(static_cast<unsigned>(x1), static_cast<unsigned>(y2)));
+        vsum q21 = ima(point2d(static_cast<unsigned>(x2), static_cast<unsigned>(y1)));
+        vsum q22 = ima(point2d(static_cast<unsigned>(x2), static_cast<unsigned>(y2)));
 
         double x2_x1 = x2 - x1;
         double y2_y1 = y2 - y1;
